requirements.txt
cffi==1.15.1
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
